+++ lib/Echidna/Campaign.hs
@@ -34,11 +34,13 @@ import System.Random (mkStdGen)
 import qualified Data.HashMap.Strict  as H
 
 import Echidna.ABI
+import Echidna.Defaults (defaultDict)
 import Echidna.Exec
 import Echidna.Solidity
+import Echidna.Solidity.Types (SolTest)
 import Echidna.Test
 import Echidna.Transaction
-import Echidna.Types (TestState(..), Campaign(..), CampaignConf(..), SolTest, GenDict, Tx, defSeed, constants, coverage, genDict, defaultDict, tests, call, rTypes)
+import Echidna.Types (TestState(..), Campaign(..), CampaignConf(..), GenDict, Tx, defSeed, constants, coverage, genDict, tests, call, rTypes)
 
 instance MonadThrow m => MonadThrow (RandT g m) where
   throwM = lift . throwM
